Refactored structure of channels related code (including graphs) to be in the Views layer
