+++ brain-tag/src/App/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.css';
 import io from 'socket.io-client';
-import ChannelContainer from '../PowerGraph/ChannelContainer';
 import Buttons from '../Views/Components/Buttons/';
+import Channels from '../Views/Components/Channels/'
 
 function App() {
 
@@ -15,7 +15,7 @@ function App() {
 		<div className="App">
 			<p className="header" > Chart View </p>
       		<div className="Charts">
-				<ChannelContainer className="chart" sampleRate='1000' intervalSize ='1000' socket={socket}  />
+				<Channels className="chart" sampleRate='1000' intervalSize ='1000' socket={socket}  />
 			</div>
 			<Buttons/>
 		</div>
